script for this modpack installer
The script is written in python 3. It needs the magic libary (It should be preinstalled I think): pip3 install magic-python. All other libaries sould be preinstalled. I tested the script under Linux (Ubuntu) and with various modpacks, but as this is a pretty long and complex code it could be that there are some unseen bugs. If there is one then I'll try to fix it.
I did not edit the other scripts (of this project).